refactor(category) : 코드 리팩토링 및 CompletedCategory Entity 수정
