+++ src/main/java/site/coach_coach/coach_coach_server/category/service/CategoryService.java
@@ -1,6 +1,7 @@
 package site.coach_coach.coach_coach_server.category.service;
 
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import lombok.RequiredArgsConstructor;
 import site.coach_coach.coach_coach_server.category.domain.Category;
@@ -8,9 +9,7 @@ import site.coach_coach.coach_coach_server.category.dto.CreateCategoryRequest;
 import site.coach_coach.coach_coach_server.category.exception.NotFoundCategoryException;
 import site.coach_coach.coach_coach_server.category.repository.CategoryRepository;
 import site.coach_coach.coach_coach_server.common.constants.ErrorMessage;
-import site.coach_coach.coach_coach_server.common.exception.AccessDeniedException;
 import site.coach_coach.coach_coach_server.routine.domain.Routine;
-import site.coach_coach.coach_coach_server.routine.exception.NoExistRoutineException;
 import site.coach_coach.coach_coach_server.routine.repository.RoutineRepository;
 import site.coach_coach.coach_coach_server.routine.service.RoutineService;
 
@@ -21,8 +20,9 @@ public class CategoryService {
 	private final RoutineService routineService;
 	private final RoutineRepository routineRepository;
 
+	@Transactional
 	public Long createCategory(CreateCategoryRequest createCategoryRequest, Long routineId, Long userIdByJwt) {
-		Routine routine = validateAccessToRoutine(routineId, userIdByJwt);
+		Routine routine = routineService.validateAccessToRoutine(routineId, userIdByJwt);
 
 		Category category = Category.builder()
 			.routine(routine)
@@ -33,8 +33,9 @@ public class CategoryService {
 		return categoryRepository.save(category).getCategoryId();
 	}
 
+	@Transactional
 	public void deleteCategory(Long routineId, Long categoryId, Long userIdByJwt) {
-		validateAccessToRoutine(routineId, userIdByJwt);
+		routineService.validateAccessToRoutine(routineId, userIdByJwt);
 
 		Boolean isExistCategory = categoryRepository.existsById(categoryId);
 		if (isExistCategory) {
@@ -42,21 +43,5 @@ public class CategoryService {
 		} else {
 			throw new NotFoundCategoryException(ErrorMessage.NOT_FOUND_CATEGORY);
 		}
-	}
-
-	public Routine validateAccessToRoutine(Long routineId, Long userIdByJwt) {
-		Routine routine = routineRepository.findById(routineId)
-			.orElseThrow(() -> new NoExistRoutineException(ErrorMessage.NOT_FOUND_ROUTINE));
-
-		if (!routine.getUserId().equals(userIdByJwt)) {
-			Long coachId = routineService.getCoachId(userIdByJwt);
-			if (routine.getCoachId() != coachId) {
-				throw new AccessDeniedException();
-			}
-		} else if (routine.getCoachId() != null) {
-			throw new AccessDeniedException();
-		}
-
-		return routine;
 	}
 }
